backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# MongoDB
motor==3.4.0

# PDF Processing
PyMuPDF==1.23.6
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
regex==2023.10.3
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
